+++ CHANGES.md
@@ -1,9 +1,11 @@
 # ChangeLog
-## v.x.x.x
+## v22.2.1
 * Fix bug with size of 'overlap' spinboxes expanding in the vertical direction
-
-## vx.x.x
 * Create environment file for development of iDVC
+* On load/creation of pointcloud display points only by default
+* Makes a few GUI bugfixes, such as fix number of points in the pointcloud displayed wrongly here and there.
+* Catches exceptions on load of files and opens message boxes
+* Fix load of TIFF files
 
 ## v22.2.0
 * Update DVC executable version to v22.0.0